requirements.txt
Django
django-ckeditor
django-js-asset
Pillow
pylint
pylint-django
pylint-plugin-utils

